fix type bug

+++ 0x1E-search_algorithms/0-linear.c
@@ -18,7 +18,7 @@ int linear_search(int *array, size_t size, int value)
 
 	for (i = 0; i < size; i++)
 	{
-		printf("value checked array[%u] = [%d]\n", (int)i, array[i]);
+		printf("value checked array[%u] = [%d]\n", (int)i, array[(int)i]);
 		if (value == array[(int)i])
 			return ((int)i);
 	}
